Add support for BCM2712/Raspberry Pi 5B
My main goal with this patch is to allow for sel4test to work on the
Raspberry Pi 5B, for this reason I have not supported as many
peripherals as there is for BCM2711.

Mainly GPIO is not supported.

My understanding is that the default UART, unlike previous BCM SoCs
supported, is not on the GPIO pins anymore.

The BCM2712 does still have support for the BCM mini UART however
it is now placed on the PCIe bus rather than being directly
connected to the SoC. The default UART for BCM2712 is instead
the PL011 so I've added support for that.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>